% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.prcomp}
\alias{as.prcomp}
\title{Convert to prcomp object}
\usage{
as.prcomp(object, x, lambda)
}
\arguments{
\item{object}{fps object}

\item{x}{data to project}

\item{lambda}{lambda value to extract}
}
\description{
Converts an fps object to a prcomp object
}

