% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cv.fps}
\alias{cv.fps}
\title{Estimate projection score by cross-validation}
\usage{
\method{cv}{fps}(object, x, K = 5, FUN = var, ...)
}
\arguments{
\item{object}{fps object}

\item{x}{Data}

\item{K}{Number of cross-validation folds}

\item{FUN}{Function to compute input matrix to fps() from subsets of x}

\item{...}{Additional arguments to fps()}
}
\description{
Estimate projection score by cross-validation
}

