% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelity.R
\name{getFidelity}
\alias{getFidelity}
\title{Retrieve Most Recent Closing Prices from Fidelity}
\usage{
getFidelity()
}
\value{
A data.table object with column Symbol, Name, NAV and Change.
}
\description{
Retrieve Most Recent Closing Prices from Fidelity
}
\details{
This is fairly raw code which may break.
}
\examples{
getFidelity()
}
