% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{plot.fps_cv}
\alias{plot.fps_cv}
\title{Plot result of cross-validation}
\usage{
\method{plot}{fps_cv}(x, ...)
}
\arguments{
\item{x}{fps_cv object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot result of cross-validation
}
