% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.svps}
\alias{plot.svps}
\title{Plot leverages}
\usage{
\method{plot}{svps}(x, type = c("row", "column"), ...)
}
\arguments{
\item{x}{svps object}

\item{type}{plot rows (default) or columns?}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function plots the regularization path
}
\examples{
data(wine)
out <- svps(scale(wine), ndim = 2)
plot(out)
}

