% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.fps}
\alias{predict.fps}
\title{Project data onto estimated subspace}
\usage{
\method{predict}{fps}(object, x, lambda, ...)
}
\arguments{
\item{object}{fps object}

\item{x}{data to project}

\item{lambda}{lambda value to extract}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Returns the projection of x onto the estimated subspace
}

