% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fps}
\alias{print.fps}
\title{Print FPS object}
\usage{
\method{print}{fps}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fps object}

\item{digits}{number of significant digits}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function prints
}
