% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fps_coef}
\alias{print.fps_coef}
\title{Print basis coefficients}
\usage{
\method{print}{fps_coef}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fps_coef object}

\item{digits}{number of significant digits}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function prints
}
