% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.svps_coef}
\alias{print.svps_coef}
\title{Print basis coefficients}
\usage{
\method{print}{svps_coef}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fps_coef object}

\item{digits}{number of significant digits}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function prints
}

