% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{projection}
\alias{projection}
\title{Extract projection matrix estimate}
\usage{
projection(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}
}
\description{
\code{projection} is a generic function for extracting a fitted projection 
matrix from a dimension reduction object
}
