% Generated by roxygen2 (4.0.1): do not edit by hand
\name{projection.svps}
\alias{projection.svps}
\title{Extract projection estimates}
\usage{
\method{projection}{svps}(object, lambda, type = c("bi", "row", "column"),
  ...)
}
\arguments{
\item{object}{fps object}

\item{lambda}{lambda value to extract}

\item{type}{row-, column-, or bi- projection?}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Returns estimated projection matrix
}

