% Generated by roxygen2 (4.0.1): do not edit by hand
\name{svps}
\alias{svps}
\title{Singular Value Projection and Selection}
\usage{
svps(x, ndim, nsol = 50L, maxblocksize = 0L, lambdaminratio = -1,
  lambdamin = -1, lambda = as.numeric(c()), maxiter = 100L,
  tolerance = 0.001, verbose = 0L)
}
\arguments{
\item{x}{Input matrix}

\item{ndim}{Target subspace dimension (can be fractional)}

\item{nsol}{Number of solutions to compute}

\item{maxblocksize}{Suggested maximum block size (rows + columns);
ignored if \code{== 0}}

\item{lambdaminratio}{Minimum value of lambda as a fraction of
the automatically determined maximum value of
lambda; ignored if \code{< 0}}

\item{lambdamin}{Minimum value of lambda; automatically determined
if \code{< 0}}

\item{lambda}{Vector of regularization parameter values}

\item{maxiter}{Maximum number of iterations for each solution}

\item{tolerance}{Convergence threshold}

\item{verbose}{Level of verbosity; silent if \code{= 0};
                      otherwise display more messages and progress
                      indicators as \code{verbose} increases}
}
\value{
An S3 object of class \code{svps} which is a list with the
        following components:
  \item{ndim}{Sum of squares (dimension) of the estimate}
  \item{lambda}{Vector containing the regularization parameters of each
                estimate.}
  \item{projection}{List containing the the (bi-)projection matrix
                    estimates}
  \item{leverage.row}{Matrix whose columns are the row leverages}
  \item{leverage.col}{Matrix whose columns are the column leverages}
  \item{L1}{Vector of the sum of absolute values of each estimate}
  \item{var.row}{}
  \item{var.col}{}
  \item{var.total}{}
  \item{niter}{a vector containing the number of ADMM iterations for each
               estimate}
}
\description{
This function computes a solution path of the Singular Value Projection
and Selection (SVPS) estimator.  It takes a data matrix \code{x} as input
and returns an object containing a list of projection matrices
estimated by SVPS over a sequence of regularization parameter values.
}
\details{
By default, a sensible minimum value of the regularization parameter
is automatically chosen so that the minimally regularized solution
is different from the ordinary SVD.

The default automatic choice of \code{lambdamin} ensures that the least
regularized estimate omits at least one row or column. The solutions
are automatically sorted in decreasing order of \code{lambda}.
}
\examples{
# Apply FPS to the standardized wine data from the UCI ML repository
data(wine)
out <- svps(scale(wine), ndim = 3)
print(out)
plot(out)
}

