% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.fps}
\alias{coef.fps}
\title{Extract basis vector estimates}
\usage{
\method{coef}{fps}(object, lambda, ...)
}
\arguments{
\item{object}{fps object}

\item{lambda}{lambda value to extract}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Returns estimated basis vectors
}

