% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.fps}
\alias{plot.fps}
\title{Plot FPS object}
\usage{
\method{plot}{fps}(x, type = c("leverage", "variance", "coherence",
  "crossleverage", "sparsity"), ...)
}
\arguments{
\item{x}{fps x}

\item{type}{type of plot}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function plots the regularization path
}
\examples{
data(wine)
out <- fps(cor(wine), ndim = 2)
plot(out)
}

