% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.fps_cv}
\alias{plot.fps_cv}
\title{Plot result of cross-validation}
\usage{
\method{plot}{fps_cv}(x, ...)
}
\arguments{
\item{x}{fps_cv object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot result of cross-validation
}

