% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.svps}
\alias{print.svps}
\title{Print SVPS object}
\usage{
\method{print}{svps}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{svps object}

\item{digits}{number of significant digits}
}
\description{
This function prints
}

