% Generated by roxygen2 (4.0.1): do not edit by hand
\name{projection.fps}
\alias{projection.fps}
\title{Extract projection estimates}
\usage{
\method{projection}{fps}(object, lambda)
}
\arguments{
\item{object}{fps object}

\item{lambda}{lambda value to extract}
}
\description{
Returns estimated projection matrix
}

