% Generated by roxygen2 (4.0.1): do not edit by hand
\name{projection}
\alias{projection}
\title{Extract projection matrix estimate}
\usage{
projection(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}
}
\description{
\code{projection} is a generic function for extracting a fitted projection
matrix from a dimension reduction object
}

