% Generated by roxygen2 (4.0.1): do not edit by hand
\name{projection}
\alias{projection}
\title{Extract projection matrix estimate}
\usage{
projection(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Extract projection matrix estimate
}

